refactor: Refactor get_latest_timestamp
