+++ callrail/callrail_service.py
@@ -5,27 +5,7 @@ from compose import compose
 
 from callrail.pipeline.interface import Pipeline
 from callrail.callrail_repo import get
-from db.bigquery import get_last_timestamp, load, update
-
-
-def _build_params_service(pipeline: Pipeline):
-    def _svc(timeframe: tuple[Optional[str], Optional[str]]) -> dict[str, str]:
-        start, end = timeframe
-        if start and end:
-            _start, _end = start, end
-        else:
-            _start = (
-                get_last_timestamp(pipeline.name, pipeline.cursor_key)
-                .date()
-                .isoformat()
-            )
-            _end = datetime.utcnow().date().isoformat()
-        return {
-            "start_date": _start,
-            "end_date": _end,
-        }
-
-    return _svc
+from db.bigquery import load, update
 
 
 def _batched_at_service(rows: list[dict[str, Any]]) -> list[dict[str, Any]]:
@@ -59,5 +39,5 @@ def pipeline_service(
         _batched_at_service,
         pipeline.transform,
         get(pipeline.uri, pipeline.res_fn),
-        _build_params_service(pipeline),
+        pipeline.params_fn(pipeline.name, pipeline.cursor_key),
     )((start, end))
