+++ test/test_units.py
@@ -6,7 +6,7 @@ from tasks import tasks_service
 
 TIME_FRAME = [
     ("auto", (None, None)),
-    # ("manual", ("2021-01-01", "2022-05-01")),
+    ("manual", ("2021-01-01", "2022-05-01")),
 ]
